+++ ocaml-lsp-server/test/e2e-new/lsp_helpers.ml
@@ -2,9 +2,9 @@ open Test.Import
 
 let change_config ~client params = Client.notification client (ChangeConfiguration params)
 
-let open_document ~client ~uri ~source =
+let open_document ~language_id ~client ~uri ~source =
   let textDocument =
-    TextDocumentItem.create ~uri ~languageId:"ocaml" ~version:0 ~text:source
+    TextDocumentItem.create ~uri ~languageId:language_id ~version:0 ~text:source
   in
   Client.notification
     client
@@ -44,28 +44,18 @@ let open_document_with_client ~prep ~path ~source client =
   in
   let uri = DocumentUri.of_path path in
   let* () = prep client in
-  open_document ~client ~uri ~source
+  open_document ~language_id:"ocaml" ~client ~uri ~source
 ;;
 
-let iter_lsp_response
+let iter_lsp_response_internal
       ?(prep = fun _ -> Fiber.return ())
       ?(path = "foo.ml")
+      ~language_id
       ~makeRequest
       ~source
       k
   =
-  let got_diagnostics = Fiber.Ivar.create () in
-  let handler =
-    Client.Handler.make
-      ~on_notification:(fun _ -> function
-         | PublishDiagnostics _ ->
-           let* diag = Fiber.Ivar.peek got_diagnostics in
-           (match diag with
-            | Some _ -> Fiber.return ()
-            | None -> Fiber.Ivar.fill got_diagnostics ())
-         | _ -> Fiber.return ())
-      ()
-  in
+  let handler = Client.Handler.make ~on_notification:(fun _ _ -> Fiber.return ()) () in
   Test.run ~handler
   @@ fun client ->
   let run_client () =
@@ -78,22 +68,46 @@ let iter_lsp_response
     in
     Client.start client (InitializeParams.create ~capabilities ())
   in
-  let run =
+  let request () =
     let* (_ : InitializeResult.t) = Client.initialized client in
     let uri = DocumentUri.of_path path in
     let* () = prep client in
-    let* () = open_document ~client ~uri ~source in
-    let+ resp =
-      let request =
-        let textDocument = TextDocumentIdentifier.create ~uri in
-        makeRequest textDocument
-      in
-      Client.request client request
+    let* () = open_document ~language_id ~client ~uri ~source in
+    let* response =
+      Fiber.collect_errors (fun () ->
+        let request =
+          let textDocument = TextDocumentIdentifier.create ~uri in
+          makeRequest textDocument
+        in
+        Client.request client request)
     in
-    k resp
+    k response
   in
-  Fiber.fork_and_join_unit run_client (fun () ->
-    run >>> Fiber.Ivar.read got_diagnostics >>> Client.stop client)
+  let shutdown () =
+    let* () = Client.request client Shutdown in
+    Client.notification client Exit
+  in
+  Fiber.fork_and_join_unit run_client (fun () -> Fiber.finalize request ~finally:shutdown)
+;;
+
+let iter_lsp_response ?prep ?path ~language_id ~makeRequest ~source k =
+  iter_lsp_response_internal ?prep ?path ~language_id ~makeRequest ~source (function
+    | Ok response ->
+      k response;
+      Fiber.return ()
+    | Error errors -> Fiber.reraise_all errors)
+;;
+
+let iter_lsp_response_result ?prep ?path ~language_id ~makeRequest ~source k =
+  iter_lsp_response_internal ?prep ?path ~language_id ~makeRequest ~source (function
+    | Ok response ->
+      k (Ok response);
+      Fiber.return ()
+    | Error [ { Exn_with_backtrace.exn = Jsonrpc.Response.Error.E error; backtrace = _ } ]
+      ->
+      k (Error error);
+      Fiber.return ()
+    | Error errors -> Fiber.reraise_all errors)
 ;;
 
 let open_document_with_diagnostics_callback